requirements.txt
ultralytics==8.2.22
pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.3
scikit-image==0.22.0
scikit-learn==1.4.1.post1